Create maestro_util for utility functions
optional flush
create maestro_util.py for utility functions
